#407 Rename plugin
