+++ strapi/src/plugins/waste-collection-days-import/admin/src/components/PluginIcon/index.tsx
