+++ strapi/src/plugins/waste-collection-days-import/admin/src/pages/App/index.tsx
