+++ strapi/src/plugins/waste-collection-days-import/admin/src/translations/en.json
