+++ strapi/src/plugins/waste-collection-days-import/admin/src/translations/fr.json
