+++ strapi/src/plugins/waste-collection-days-import/admin/src/utils/getTrad.ts
