+++ strapi/src/plugins/waste-collection-days-import/custom.d.ts
