+++ strapi/src/plugins/waste-collection-days-import/server/register.ts
